requirements.txt
flask
pytest
python-dotenv
pytest-cov
mysql-connector-python
